Extract `FeedViewController` test helpers in their own file
